Update 15.10.12 Addition to #34
The core, the loader and the saver should now fully support this, however this could not yet be fully tested, so I cannot yet guarantee no bugs may exist.
Tomorrow I am going to see if I can adept the editor to this feature, which will be one hell of a job (so wish me luck) as I need to alter the layout and also the canvas for that, and the latter in particular is a very sensitive thing.
Of course, this update will also enhance the Star Story project, though that project will also need to dig through for full support. I actually should have done this already for the Yaqirpa dungeon, but... I was a bit lazy at the time (well, lazy is the wrong word, I just got tired of that place as it took ages to develop that place, which is quite normal for the first dungeon, but I guess I get the bill now for not doing it right away.

Of course, I am not going to close #34 yet, once the editor fully supports this I am going to close this.

See you soon